+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Initial release of Oragono!
 ### Added
 * Added YAML config file format.
 * Added native SSL/TLS support (thanks to @edmand).
+* Added ability to generate certificates from the command line.
 * We now advertise the [`RPL_ISUPPORT`](http://modern.ircdocs.horse/#rplisupport-005) numeric.
 * Parse new mode change syntax commonly used these days (i.e. `+h-ov dan dan dan`).